Scroll view in screening info pop-up
Aded a scroll view to ScreeningDialogController to contain the go-to-screening buttons and labels.
Unified the filling of the documentview (containing controls for all screenings involved) between ScreeningDialogController and FilmInfoDialogController, and put the code into abstract class GoToScreeningControllerView.
Refactored the creation of Friend Controls (Friend Rating combo boxes and Friend Attendance checkboxes) as to simplify the placement of the controls on their view.
Cloned the My Rating combo box from the one defined in Xcode and dispoded the original, I could'n get the Friend Rating combo boxes clones look exactly like it.