Add tests
* models.py:
* models.py: Support temp dir for testing.

* tests.py: Fill more attributes in test films.
New test to demonstrate that blank reviewer can go into the database.
Repair test that was broken by the changed tools.

* views.py: Recognize absent subsection being None.

* loader_forms.py: Convert any blank string representing subsection ot
reviewer to None.

* reviewers.html: Fix typo.

* tests.py: Support temp dir for testing.
New tests to demonstrate that blank reviewer will go into the database
as None when loading.
Rewrite tests that were dependant of the loader project.

* views.py: Fix filtering of unsaved model objects, will be deprecated
in Django version 5.0.

* manage.py: Use explicit python runner.