Follow-up code review
* screening_status_getter.py: Add comment to help understanding how
"items()" of a dictionary of dictonaries works.

* tests.py: Add test to help understanding how "items()" of a
dictionary of dictonaries works.

* parse_idfa_html.py: Add docstring.
Move function to a more logical place.
New method "broadcast()" that writes a message to the console and the
debug text.
Remove superfluous derived method.
Integrate "if" statement into existing "match".
Move compilation of regex to class attributes.
Add small function as to keep "handle_xyz()" methods as simple as
possible.
Simplify code.
Remove superfluous comments.

* parse_tools.py:
* parse_iffr_html.py:
* application_tools.py: New method "broadcast()" that writes a message
to the console and the debug text.

* planner_interface.py: Disactivate XML dumper.